Separate autobem4_to_glm from buildings module
