+++ tools/autobem4_to_glm.py
@@ -1,0 +1,35 @@
+"""Generate GLM model from AutoBEM4 buildings data"""
+import sys
+import datetime as dt
+from typing import TypeVar
+
+def write(buildings:TypeVar('Buildings'),filename:str):
+
+    buildings.data.set_index('id',inplace=True)
+    with open(filename,"w") as fh:
+        fh.write(f"""// generated by {sys.argv} on {dt.datetime.now()}
+class building
+{{
+char8 climate;
+int16 year;
+char32 centroid;
+char256 footprint;
+double height[m];
+double ground_area[m^2];
+char32 code;
+char32 IECC;
+bool mixed;
+char32 type;
+double windows[pu];
+int16 floors;
+double floor_area[m^2];
+}}
+""")
+        for name,data in buildings.data.iterrows():
+            properties = "\n".join([f"""    {x} "{y}";""" for x,y in zip(data.index,data.values)]);
+            fh.write(f"""object building
+{{
+name "B_{name}";
+{properties}
+}}
+""")
