+++ tools/buildings.py
@@ -83,6 +83,7 @@ import gridlabd.geodata_address as address
 import gridlabd.nsrdb_weather as nw
 from typing import TypeVar
 import contextlib
+import autobem4_to_glm
 
 class BuildingsError(app.ApplicationError):
     """Buildings exception"""
@@ -182,36 +183,6 @@ class Buildings:
 
         self.data["address"] = address.apply(self.data)
 
-    def _to_glm(self,filename):
-
-        self.data.set_index('id',inplace=True)
-        with open(filename,"w") as fh:
-            fh.write(f"""// generated by {sys.argv} on {dt.datetime.now()}
-class building
-{{
-    char8 climate;
-    int16 year;
-    char32 centroid;
-    char256 footprint;
-    double height[m];
-    double ground_area[m^2];
-    char32 code;
-    char32 IECC;
-    bool mixed;
-    char32 type;
-    double windows[pu];
-    int16 floors;
-    double floor_area[m^2];
-}}
-""")
-            for name,data in self.data.iterrows():
-                properties = "\n".join([f"""    {x} "{y}";""" for x,y in zip(data.index,data.values)]);
-                fh.write(f"""object building
-{{
-    name "B_{name}";
-{properties}
-}}
-""")
     
 def main(argv:list[str]) -> int:
     """Main routine
@@ -302,7 +273,7 @@ def main(argv:list[str]) -> int:
         # output to GLM
         elif isinstance(OUTPUT,str) and OUTPUT.endswith(".glm"):
 
-            result._to_glm(OUTPUT)
+            autobem4_to_glm.write(result,OUTPUT)
 
         # unsupported output format
         else:
@@ -313,5 +284,7 @@ def main(argv:list[str]) -> int:
     return app.E_OK
 
 if __name__ == "__main__":
+
+    main("buildings -C=US/NH/Sullivan -o=/tmp/test.glm".split())
 
     app.run(main)
